feat: Append the target commitish to the version in the manifest
The version number in the manifest should match the version reported in
the first line from "nvim --version," but unfortunately we can't get the
number of commits since the last stable release unless we keep a local
copy of the git repo. Instead we just fetch the nightly release's
target commitish and include that in the version.