fix: Fixed checksum check
Previously there was a checksum file for each artifact of a release,
however for some reason it was changed to a single shasum.txt file that
contains the checksum for all artifacts. This update fetches that file
instead and extracts the checksum just for the Windows x64 zip file,
which is then injected into the generated manifest.